requirements.txt
flask[async]
requests-async
python-gitlab
python-jenkins